Initial Commit v0.1
v0.1 - Just a basic run of the program to get the idea of the formula
and one very simple way of achieving the wanted result.
This will later be almost completely changed to try and adopt ctime and
run off of dates and time rather than a simple whole number based
formula.